Initial commit
This commit has some hard-coded paths on my computer. That is easy
enough to fix. Howvever, it has a lot of cruft because the
sheets shown to faculty did not include Spire IDs or email addresses.
I suggest we include email addresses in the future, which are enough
to disambiguate applicants. That will let us clean up a lot of the
code.